pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "eorcalc"
version = "0.1.1"
authors = [
    {name = "Hajime Hinata", email = "onmyojiflow@gmail.com"},
]
description = "GPU-accelerated reionization field calculator with power spectrum analysis for the Epoch of Reionization"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Astronomy",
    "Topic :: Scientific/Engineering :: Physics",
]
dependencies = [
    "numpy>=1.20.0",
    "scipy>=1.6.0",
    "pandas>=1.3.0",
    "astropy>=4.0",
    "filelock>=3.0.0",
    "camb>=1.3.0",
    "massfunc>=0.0.10",
    "cosfunc>=0.1.0",
    "xxiop>=0.1.0",
    "cupy-cuda12x>=12.0.0",
]
keywords = [
    "astrophysics",
    "cosmology", 
    "ionization",
    "power spectrum",
    "diffusion",
    "reionization",
    "EoR",
    "GPU",
    "ionization field",
    "optical depth",
]
license = {text = "MIT"}

[project.urls]
Homepage = "https://github.com/SOYONAOC/EoRCaLC"
Repository = "https://github.com/SOYONAOC/EoRCaLC"
"Bug Reports" = "https://github.com/SOYONAOC/EoRCaLC/issues"
Documentation = "https://github.com/SOYONAOC/EoRCaLC#readme"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=3.0",
    "black>=22.0",
    "flake8>=4.0",
    "mypy>=0.950",
    "ipython>=8.0",
    "jupyter>=1.0",
]
test = [
    "pytest>=7.0",
    "pytest-cov>=3.0",
]
gpu-cuda11 = [
    "cupy-cuda11x>=11.0.0",
]

[tool.setuptools]
packages = ["eorcalc"]

[tool.setuptools.package-dir]
eorcalc = "eorcalc"

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | __pycache__
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=eorcalc",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = [
    "camb.*",
    "massfunc.*",
    "scipy.*",
    "astropy.*",
    "filelock.*",
    "cupy.*",
    "cosfunc.*",
    "xxiop.*",
    "pandas.*",
]
ignore_missing_imports = true
